Added support for 8192 and 16384 F32 FFTs.
It has an impact on other FFTs since they may use the same bit reverse
table with the bigger uint32 offset.

The helium code for bit reverse of int16 cannot easily support uint32
offset and will have to be changed.
So, for q15 and f16, the bit reverse is scalar instead of using Helium
instructions.
It is temporary.

8192 and 16384 not yet tested. It builds, tests for size <= 4096 are
passing (scalar, helium and all datatypes).

So nothing is broken.

+++ Include/dsp/transform_functions_f16.h
@@ -80,7 +80,7 @@ extern "C"
   {
           uint16_t fftLen;                   /**< length of the FFT. */
     const float16_t *pTwiddle;         /**< points to the Twiddle factor table. */
-    const uint16_t *pBitRevTable;      /**< points to the bit reversal table. */
+    const uint32_t *pBitRevTable;      /**< points to the bit reversal table. */
           uint16_t bitRevLength;             /**< bit reversal table length. */
 #if defined(ARM_MATH_MVEF) && !defined(ARM_MATH_AUTOVECTORIZE)
    const uint32_t *rearranged_twiddle_tab_stride1_arr;        /**< Per stage reordered twiddle pointer (offset 1) */                                                       \
