+++ Source/TransformFunctions/arm_cfft_init_f16.c
@@ -94,7 +94,7 @@ arm_status arm_cfft_init_##LEN##_f16(                                  \
                                                                        \
     /*  Initialise the bit reversal table modifier */                  \
     S->bitRevLength = ARMBITREVINDEXTABLE_FIXED_##LEN##_TABLE_LENGTH;  \
-    S->pBitRevTable = (uint16_t *)armBitRevIndexTable_fixed_##LEN;     \
+    S->pBitRevTable = (uint32_t *)armBitRevIndexTable_fixed_##LEN;     \
     S->pTwiddle = (float16_t *)twiddleCoefF16_##LEN;                   \
     status=arm_cfft_radix4by2_rearrange_twiddles_##LENTWIDDLE##_f16(S);\
                                                                        \
